speedup data reading, fix last byte loss
Compass reset data lines without waiting for the disk, without waiting for us to raise the NDAC. This is a violation of the standard, but there's nothing we can do about it. We just need to read the data as quickly as possible before it's lost.

It's a dangerous situation, but there's no solution yet.