connect only requested disks
Previously, to debug the emulator with a real device on bus, it was necessary to comment some addresses in KnownDevices.

Also emulator incorrectly handled "uninserted" image, which cause issues with laptop.

There is no measurable difference in speed between the old approach and the new one, but this makes further improvements much easier.