rewrite listener gpib handshake from scratch
After reverse engineering the 2101 floppy drive firmware, it turned out
that the floppy drive behaves differently on the bus.

I decided not to follow the standard, but to transfer the logic from the
original peripheral.

This doesn’t fix anything explicitly, but now I’m confident in the
correctness of the handshakes and command handling.